requirements/base.txt
-i https://pypi.org/simple
anyio==4.4.0; python_version >= '3.8'
asgiref==3.8.1; python_version >= '3.8'
bcrypt==4.1.3; python_version >= '3.7'
certifi==2024.6.2; python_version >= '3.6'
cffi==1.16.0; platform_python_implementation != 'PyPy'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
cryptography==42.0.8; python_version >= '3.7'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
distlib==0.3.8
dj-database-url==2.2.0
django==5.0.6; python_version >= '3.10'
django-allauth==0.63.3; python_version >= '3.7'
django-debug-toolbar==4.4.2; python_version >= '3.8'
django-environ==0.11.2; python_version >= '3.6' and python_version < '4'
django-extensions==3.2.3; python_version >= '3.6'
django-filter==24.2; python_version >= '3.8'
django-htmx==1.17.3; python_version >= '3.8'
django-widget-tweaks==1.5.0; python_version >= '3.8'
ecdsa==0.19.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
factory-boy==3.3.0; python_version >= '3.7'
filelock==3.15.1; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
httptools==0.6.1
idna==3.7; python_version >= '3.5'
json-log-formatter==1.0; python_version >= '3.6'
oauthlib==3.2.2; python_version >= '3.6'
packaging==24.1; python_version >= '3.8'
pathspec==0.12.1; python_version >= '3.8'
pipenv==2024.0.1; python_version >= '3.8'
platformdirs==4.2.2; python_version >= '3.8'
psycopg2-binary==2.9.9; python_version >= '3.7'
pyasn1==0.6.0; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1
python-jose==3.3.0
python3-openid==3.2.0
pyyaml==6.0.1
redis==5.0.6; python_version >= '3.7'
requests==2.32.3; python_version >= '3.8'
requests-oauthlib==2.0.0; python_version >= '3.4'
rsa==4.9; python_version >= '3.6' and python_version < '4'
sentry-sdk==2.5.1; python_version >= '3.6'
setuptools==70.0.0; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
sqlparse==0.5.0; python_version >= '3.8'
tomli==2.0.1; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2024.1; python_version >= '2'
urllib3==2.2.1; python_version >= '3.8'
uvicorn[standard]==0.30.1; python_version >= '3.8'
uvloop==0.19.0
uwsgi==2.0.26
virtualenv==20.26.2; python_version >= '3.7'
watchfiles==0.22.0
websockets==12.0
